+++ docs/changelog.md
@@ -10,6 +10,7 @@
 - Implemented a limit that limits how many instructions can be run at max (currently 1 million)
     - This ensures that the program will not freeze when an infinite loop is executed
     - Added option `--disable-instruction-limit` with which this limit can be circumvented
+- Keybind hint for `n` is now dynamically set to `Run to end [n]` when no breakpoint is set or to `Next breakpoint [n]` when at least one breakpoint is set
 
 ## v1.0.1